Random spawning finished/progressed
Cubes spawn randomly
As well as the identifiers.
Next step is to make at least a cube spawn the same color as the identifier in order for the player to always have a cube to click.
Identifiers need changes; is not supposed to spawn like cubes needs a different spawning method.