Cube and Identifier spawning fixed
Now the at least one cube is as the same color as the identifier.
-The cubes now are destroyed when they touch the identifier.
-Next point is to destroy cubes as they are being clicked or tapped on.
-Next point is to spawn cubes and Ident half a second after they are destroyed (by being tapped on or colliding with eachother). I MORE THOUGHTS  ON THIS.
Identifier should only spawn at the same time as cubes and cubes should spawn half a second later or after a time to give the player time to get ready for the next wave of cubes. It could be a fixed 1 second time after they are destroyed but applying the challenging changes to the cubes and identifiers.