Multiplier Text Finished with animation & Cube Spawning times and window time half way implemented
Multiplier Text: is finished with animation. It works as it should. Later on I need to add better text quality and fire animations with their own color etc. Is all aestethic.
Window respawn time: Now cubes won't spawn unless the first wave of cubes is destroyed this allows for two waves of cubes not to spawn before time. Needs more work but is almost finished.