Milestone Reached!: Working on levels Now!
Spawning and control of number of cubes is complete. Now I can decide depending on level how many cubes we can spawn and which colors I want to spawn!.
Before I continue working on the levels I need to find out how to spawn more than one cube of the same color wihtout spawning 3 of the same just 2. Later on Spawn 3 or more but that is the idea.