MainMenu in progress
Main Menu buttons are done I learned the functionalities and how to set up the transition between seens. Now if we click the play button we go into the game to play. I  need to set the volume options as well as the store and scoreboard. Menu should be done this week with the exception of scoreboard since that will need a database online.