Losing Screen
I have finished the losing screen functionality. It's complete. If i click main menu I got to the main menu. If I click Watch video "it will show a video" and then give the user a life and start the game from where the player as playing. If I click play again the game restarts with all its values and I can play the game form the beginning