Music visualizaton in progress
It has been successfull I am now make the cubes scale depending on the beat of the music and they visualize the music I need to work more on this and then transition to particles to see how particles could react with this or how to do it. This is probably gonna be one of the hardest parts with particles as i dont know if it would be with the emission of particles and size of particles all at the same time.