Working on sphere multiplier
Work on the spheres' multiplier, add the lives reduction, set up the levels, and once that is done, I have to set up the transition between the cubes and the spheres level by disableling the cubes after the user reaches the level. then work on the speed in the spheres. after this the game is finished and music as well as some graphics should be added.