Tap and collision for spheres finished
I added the correct live counter decreaser and increaser for when the user taps the correct color, the wrong color, and when the matching identifier and colored sphere collide and decreases the lives counter. Next up is create the transition between cubes and colors. After that fixed the sphere identifier collider mesh.