Cubes destruction animation finished.
There is a bug with the blue cube. fix it!
If the first cube is blue after pressing play it will not detect it as a valid cube and will count towards the points but it will not reset for the next wave of cubes making the player lose lives. fix it!